requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
alembic==1.13.1

pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

passlib[bcrypt]==1.7.4
bcrypt==4.1.3

python-jose==3.3.0

python-multipart==0.0.9

requests==2.32.3